:truck: Move Validator to validator.go
