+++ interfaces/handler/router.go
@@ -3,21 +3,17 @@ package handler
 import (
 	"log"
 
-	"github.com/go-playground/validator/v10"
-
 	"github.com/labstack/echo/v4"
 	"github.com/labstack/echo/v4/middleware"
 )
 
 func Setup(e *echo.Echo, api API) {
-
-	v := validator.New()
-	if err := v.RegisterValidation("is-uuid", IsValidUUID); err != nil {
+	// Setup validator
+	v, err := newValidator()
+	if err != nil {
 		log.Fatal(err)
 	}
-	e.Validator = &Validator{
-		validator: v,
-	}
+	e.Validator = v
 
 	// Middleware
 	e.Use(middleware.Logger())
@@ -150,12 +146,4 @@ func Setup(e *echo.Echo, api API) {
 		}
 	}
 
-}
-
-type Validator struct {
-	validator *validator.Validate
-}
-
-func (v *Validator) Validate(i interface{}) error {
-	return v.validator.Struct(i)
 }
