+++ interfaces/handler/validator.go
@@ -24,7 +24,24 @@ func (c *Context) BindAndValidate(i interface{}) error {
 	return nil
 }
 
-func IsValidUUID(fl validator.FieldLevel) bool {
+type Validator struct {
+	validator *validator.Validate
+}
+
+func newValidator() (*Validator, error) {
+	v := validator.New()
+	if err := v.RegisterValidation("is-uuid", isValidUUID); err != nil {
+		return nil, err
+	}
+
+	return &Validator{v}, nil
+}
+
+func (v *Validator) Validate(i interface{}) error {
+	return v.validator.Struct(i)
+}
+
+func isValidUUID(fl validator.FieldLevel) bool {
 	id, ok := fl.Field().Interface().(uuid.UUID)
 	return ok && id != uuid.Nil
 }
